sec commit

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -9,7 +9,7 @@ int main(void)
 	char string[] = "_putchar";
 	int i = 0;
 
-	while (i < strlen(string))
+	while (i < (int)strlen(string))
 	{
 		_putchar(string[i]);
 	}
